refactor: extract the default virtuoso components and message renderer
