+++ src/components/MessageList/__tests__/VirtualizedMessageListComponents.test.js
@@ -1,0 +1,377 @@
+import React from 'react';
+import '@testing-library/jest-dom';
+import { render, screen } from '@testing-library/react';
+import {
+  EmptyPlaceholder,
+  Footer,
+  Header,
+  Item,
+  messageRenderer,
+} from '../VirtualizedMessageListComponents';
+import {
+  generateChannel,
+  generateMessage,
+  generateUser,
+  getTestClientWithUser,
+} from '../../../mock-builders';
+import {
+  ChannelActionProvider,
+  ChannelStateProvider,
+  ChatProvider,
+  ComponentProvider,
+} from '../../../context';
+import { MessageSimple } from '../../Message';
+
+const prependOffset = 0;
+const user1 = generateUser();
+const user2 = generateUser();
+let client;
+let channel;
+
+const PREPEND_OFFSET = 10 ** 7;
+
+const Wrapper = ({ children, componentContext = {} }) => (
+  <ChatProvider value={{ client }}>
+    <ChannelStateProvider value={{ channel }}>
+      <ChannelActionProvider value={{ addNotification: jest.fn() }}>
+        <ComponentProvider value={componentContext}>{children}</ComponentProvider>
+      </ChannelActionProvider>
+    </ChannelStateProvider>
+  </ChatProvider>
+);
+
+const renderElements = (children, componentContext) =>
+  render(<Wrapper componentContext={componentContext}>{children}</Wrapper>);
+
+describe('VirtualizedMessageComponents', () => {
+  describe('Item', function () {
+    const processedMessages = [generateMessage()];
+    const withVirtualMessageClasses = { virtualMessage: 'XXX' };
+    const withMessageGroupStyles = { [processedMessages[0].id]: 'single' };
+    const withoutVirtualMessageClasses = undefined;
+    const withoutMessageGroupStyles = {};
+
+    it.each([
+      ['with', 'without', withVirtualMessageClasses, withoutMessageGroupStyles],
+      ['without', 'without', withoutVirtualMessageClasses, withoutMessageGroupStyles],
+      ['without', 'with', withoutVirtualMessageClasses, withMessageGroupStyles],
+      ['with', 'with', withVirtualMessageClasses, withMessageGroupStyles],
+    ])(
+      'should render wrapper %s custom classes %s group styles',
+      (_, __, customClasses, messageGroupStyles) => {
+        const props = {
+          'data-item-index': PREPEND_OFFSET,
+        };
+        const virtuosoContext = {
+          customClasses,
+          messageGroupStyles,
+          numItemsPrepended: 0,
+          processedMessages,
+        };
+
+        const { container } = renderElements(<Item context={virtuosoContext} {...props} />);
+        expect(container).toMatchSnapshot();
+      },
+    );
+  });
+
+  describe('Header', () => {
+    const head = <div>Custom head</div>;
+    const CustomLoadingIndicator = () => <div>Custom Loading Indicator</div>;
+    it('should render empty div in Header when not loading more messages', () => {
+      const { container } = renderElements(<Header />);
+      expect(container).toMatchInlineSnapshot(`<div />`);
+    });
+
+    it('should render LoadingIndicator in Header when loading more messages', () => {
+      const context = { loadingMore: true };
+      const { container } = renderElements(<Header context={context} />);
+      expect(container).toMatchSnapshot();
+    });
+
+    it('should render custom LoadingIndicator in Header when loading more messages', () => {
+      const componentContext = { LoadingIndicator: CustomLoadingIndicator };
+      const context = { loadingMore: true };
+      const { container } = renderElements(<Header context={context} />, componentContext);
+      expect(container).toMatchInlineSnapshot(`
+        <div>
+          <div
+            class="str-chat__virtual-list__loading"
+          >
+            <div>
+              Custom Loading Indicator
+            </div>
+          </div>
+        </div>
+      `);
+    });
+
+    it('should not render custom LoadingIndicator in Header when not loading more messages', () => {
+      const componentContext = { LoadingIndicator: CustomLoadingIndicator };
+      const { container } = renderElements(<Header />, componentContext);
+      expect(container).toMatchInlineSnapshot(`<div />`);
+    });
+
+    // FIXME: this is a crazy pattern of having to set LoadingIndicator to null so that additionalVirtuosoProps.head can be rendered.
+    it('should not render custom head in Header when loading more messages, but the LoadingIndicator', () => {
+      const context = { head, loadingMore: true };
+      const { container } = renderElements(<Header context={context} />);
+      expect(container).toMatchSnapshot();
+    });
+
+    // FIXME: this is a crazy pattern of having to set LoadingIndicator to null so that additionalVirtuosoProps.head can be rendered.
+    it('should render custom head in Header when LoadingIndicator in component context is set to null', () => {
+      const componentContext = {
+        LoadingIndicator: null,
+      };
+      const context = { head, loadingMore: true };
+      const { container } = renderElements(<Header context={context} />, componentContext);
+      expect(container).toMatchInlineSnapshot(`
+        <div>
+          <div>
+            Custom head
+          </div>
+        </div>
+      `);
+    });
+
+    it('should not render custom head in Header when not loading more messages', () => {
+      const context = { head };
+      const { container } = renderElements(<Header context={context} />);
+      expect(container).toMatchInlineSnapshot(`<div />`);
+    });
+
+    it('should render custom LoadingIndicator instead of head when loading more', () => {
+      const componentContext = { LoadingIndicator: CustomLoadingIndicator };
+      const context = { head, loadingMore: true };
+      const { container } = renderElements(<Header context={context} />, componentContext);
+      expect(container).toMatchInlineSnapshot(`
+        <div>
+          <div
+            class="str-chat__virtual-list__loading"
+          >
+            <div>
+              Custom Loading Indicator
+            </div>
+          </div>
+        </div>
+      `);
+    });
+  });
+
+  describe('EmptyPlaceholder', () => {
+    const EmptyStateIndicator = ({ listType }) => (
+      <div data-listType={listType}>Custom EmptyStateIndicator</div>
+    );
+    const NullEmptyStateIndicator = null;
+    const componentContext = { EmptyStateIndicator };
+    it('should render for main message list by default', () => {
+      const { container } = renderElements(<EmptyPlaceholder />);
+      expect(container).toMatchSnapshot();
+    });
+
+    it('should render empty for thread by default', () => {
+      const { container } = renderElements(<EmptyPlaceholder context={{ threadList: true }} />);
+      expect(container).toMatchInlineSnapshot(`<div />`);
+    });
+    it('should render custom EmptyStateIndicator for main message list', () => {
+      const { container } = renderElements(<EmptyPlaceholder />, componentContext);
+      expect(container).toMatchSnapshot();
+    });
+
+    it('should render custom EmptyStateIndicator for thread', () => {
+      const { container } = renderElements(
+        <EmptyPlaceholder context={{ threadList: true }} />,
+        componentContext,
+      );
+      expect(container).toMatchSnapshot();
+    });
+
+    it('should render empty if EmptyStateIndicator nullified', () => {
+      const componentContext = { EmptyStateIndicator: NullEmptyStateIndicator };
+      const { container } = renderElements(<EmptyPlaceholder />, componentContext);
+      expect(container).toMatchInlineSnapshot(`<div />`);
+    });
+
+    it('should render empty in thread if EmptyStateIndicator nullified', () => {
+      const componentContext = { EmptyStateIndicator: NullEmptyStateIndicator };
+      const { container } = renderElements(
+        <EmptyPlaceholder context={{ threadList: true }} />,
+        componentContext,
+      );
+      expect(container).toMatchInlineSnapshot(`<div />`);
+    });
+  });
+
+  describe('Footer', () => {
+    it('should render nothing in Footer by default', () => {
+      const { container } = renderElements(<Footer />);
+      expect(container).toMatchInlineSnapshot(`<div />`);
+    });
+    it('should render custom TypingIndicator in Footer', () => {
+      const TypingIndicator = () => <div>Custom TypingIndicator</div>;
+      const componentContext = { TypingIndicator };
+      const { container } = renderElements(<Footer />, componentContext);
+      expect(container).toMatchInlineSnapshot(`
+        <div>
+          <div>
+            Custom TypingIndicator
+          </div>
+        </div>
+      `);
+    });
+  });
+
+  describe('messageRenderer', () => {
+    const virtuosoIndex = PREPEND_OFFSET;
+    const numItemsPrepended = 0;
+    beforeAll(async () => {
+      client = await getTestClientWithUser();
+      const channelData = generateChannel();
+      channel = client.channel(channelData.channel.type, channelData.channel.id, channelData);
+    });
+
+    it('should allow to execute custom item rendering logic instead of the default', () => {
+      const customMessageRenderer = jest.fn();
+      const virtuosoContext = {
+        customMessageRenderer,
+        numItemsPrepended,
+        processedMessages: [generateMessage()],
+      };
+      messageRenderer(virtuosoIndex, undefined, virtuosoContext);
+      expect(customMessageRenderer).toHaveBeenCalledWith(
+        expect.arrayContaining(virtuosoContext.processedMessages),
+        0,
+      );
+    });
+
+    describe('default item rendering logic', () => {
+      it('should render nothing if MessageSystem component is undefined for system messages', () => {
+        const virtuosoContext = {
+          numItemsPrepended,
+          processedMessages: [generateMessage({ type: 'system' })],
+        };
+
+        const { container } = render(messageRenderer(virtuosoIndex, undefined, virtuosoContext));
+        expect(container).toMatchInlineSnapshot(`<div />`);
+      });
+
+      it('should render MessageSystem component for system messages', () => {
+        const text = 'MessageSystem';
+        const MessageSystem = () => <>{text}</>;
+        const virtuosoContext = {
+          MessageSystem,
+          numItemsPrepended,
+          processedMessages: [generateMessage({ type: 'system' })],
+        };
+
+        render(messageRenderer(virtuosoIndex, undefined, virtuosoContext));
+        expect(screen.getByText(text)).toBeInTheDocument();
+      });
+
+      it('should render nothing if DateSeparator component is undefined for custom message type date', () => {
+        const virtuosoContext = {
+          numItemsPrepended,
+          processedMessages: [generateMessage({ customType: 'message.date', date: new Date() })],
+        };
+        const { container } = render(messageRenderer(virtuosoIndex, undefined, virtuosoContext));
+        expect(container).toMatchInlineSnapshot(`<div />`);
+      });
+
+      it('should render DateSeparator component for custom message type date', () => {
+        const text = 'DateSeparator';
+        const DateSeparator = () => <div>{text}</div>;
+        const virtuosoContext = {
+          DateSeparator,
+          numItemsPrepended,
+          processedMessages: [generateMessage({ customType: 'message.date', date: new Date() })],
+        };
+        render(messageRenderer(virtuosoIndex, undefined, virtuosoContext));
+        expect(screen.getByText(text)).toBeInTheDocument();
+      });
+
+      it('should render empty div when trying to render message at non-existent index', () => {
+        const virtuosoIndex = PREPEND_OFFSET - 1;
+        const virtuosoContext = {
+          numItemsPrepended,
+          processedMessages: [generateMessage()],
+        };
+
+        const { container } = render(messageRenderer(virtuosoIndex, undefined, virtuosoContext));
+        expect(container).toMatchInlineSnapshot(`
+                  <div>
+                    <div
+                      style="height: 1px;"
+                    />
+                  </div>
+              `);
+      });
+
+      it.each([
+        ['not ', 'by default', 'not ', false],
+        ['', '', '', true],
+      ])(
+        'should %sgroup messages %s and mark the first and the last group message',
+        (_, __, ___, shouldGroupByUser) => {
+          const virtuosoRef = { current: {} };
+          const user1MessageGroup = [
+            generateMessage({ user: user1 }),
+            generateMessage({ user: user1 }),
+            generateMessage({ user: user1 }),
+            generateMessage({ user: user1 }),
+          ];
+          const processedMessages = [
+            generateMessage({ user: user2 }),
+            ...user1MessageGroup,
+            generateMessage({ user: user2 }),
+          ];
+
+          const { container } = renderElements(
+            <>
+              {processedMessages.map((_, numItemsPrepended) => {
+                const virtuosoContext = {
+                  Message: MessageSimple,
+                  numItemsPrepended,
+                  ownMessagesReadByOthers: {},
+                  prependOffset,
+                  processedMessages,
+                  shouldGroupByUser,
+                  virtuosoRef,
+                };
+                return (
+                  <div key={numItemsPrepended}>
+                    {messageRenderer(virtuosoIndex, undefined, virtuosoContext)}
+                  </div>
+                );
+              })}
+            </>,
+          );
+          const messageElements = container.getElementsByClassName(
+            'str-chat__message str-chat__message-simple',
+          );
+
+          const firstGroupItemClass = 'str-chat__virtual-message__wrapper--first';
+          const lastGroupItemClass = 'str-chat__virtual-message__wrapper--end';
+          expect(
+            container.getElementsByClassName('str-chat__virtual-message__wrapper--group'),
+          ).toHaveLength(shouldGroupByUser ? user1MessageGroup.length - 1 : 0);
+
+          expect(container.getElementsByClassName(firstGroupItemClass)).toHaveLength(
+            shouldGroupByUser ? 3 : 0,
+          );
+          expect(container.getElementsByClassName(lastGroupItemClass)).toHaveLength(
+            shouldGroupByUser ? 3 : 0,
+          );
+          if (shouldGroupByUser) {
+            expect(messageElements[0]).toHaveClass(firstGroupItemClass);
+            expect(messageElements[0]).toHaveClass(lastGroupItemClass);
+            expect(messageElements[1]).toHaveClass(firstGroupItemClass);
+            expect(messageElements[processedMessages.length - 2]).toHaveClass(lastGroupItemClass);
+            expect(messageElements[processedMessages.length - 1]).toHaveClass(firstGroupItemClass);
+            expect(messageElements[processedMessages.length - 1]).toHaveClass(lastGroupItemClass);
+          }
+        },
+      );
+    });
+  });
+});
