+++ src/components/MessageList/hooks/VirtualizedMessageList/index.ts
@@ -4,4 +4,3 @@ export * from './useNewMessageNotification';
 export * from './usePrependMessagesCount';
 export * from './useScrollToBottomOnNewMessage';
 export * from './useShouldForceScrollToBottom';
-export * from './useVirtualizedMessageListComponents';
